fastboot: Use size_t for send_packet argument Android Bringup: See http://go/android-fw-sync
Use size_t type for packet length argument in send_packet function. This
function is part of the FastbootOps API.

BRANCH=main
BUG=b:370465224
TEST=Compile FW

(cherry picked from commit ef9acfb029ab075fabe7c3c239ce0ab727c8627e)

Original-Signed-off-by: Tomasz Michalec <tmichalec@google.com>
Original-Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/6097461
Original-Reviewed-by: Julius Werner <jwerner@chromium.org>
Original-Reviewed-by: Jakub Czapiga <czapiga@google.com>
GitOrigin-RevId: ef9acfb029ab075fabe7c3c239ce0ab727c8627e
Cr-Build-Id: 8726272140783505793
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8726272140783505793
Copybot-Job-Name: android-main-depthcharge-copybot-downstream
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/6158448
Commit-Queue: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Bot-Commit: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
(cherry picked from commit bce105210f1c710105621826bf9171ebcf24f795)
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/6167268
Tested-by: Nick Vaccaro <nvaccaro@google.com>
Commit-Queue: Nick Vaccaro <nvaccaro@google.com>

+++ src/fastboot/fastboot.h
@@ -75,7 +75,7 @@ struct FastbootOps {
 	 */
 	void (*poll)(struct FastbootOps *fb);
 	/* Send fastboot response. This function is required. */
-	int (*send_packet)(struct FastbootOps *fb, void *buf, uint16_t len);
+	int (*send_packet)(struct FastbootOps *fb, void *buf, size_t len);
 	/* Called on fastboot_reset_session(). This function is optional. */
 	void (*reset)(struct FastbootOps *fb);
 	/*
